+++ sdk/resourcemanager/dataprotection/armdataprotection/zz_generated_models.go
@@ -1317,7 +1317,7 @@ type InnerError struct {
 // ItemLevelRestoreCriteriaClassification provides polymorphic access to related types.
 // Call the interface's GetItemLevelRestoreCriteria() method to access the common type.
 // Use a type switch to determine the concrete type.  The possible types are:
-// - *ItemLevelRestoreCriteria, *RangeBasedItemLevelRestoreCriteria
+// - *ItemLevelRestoreCriteria, *KubernetesPVRestoreCriteria, *KubernetesStorageClassRestoreCriteria, *RangeBasedItemLevelRestoreCriteria
 type ItemLevelRestoreCriteriaClassification interface {
 	// GetItemLevelRestoreCriteria returns the ItemLevelRestoreCriteria content of the underlying type.
 	GetItemLevelRestoreCriteria() *ItemLevelRestoreCriteria
@@ -1403,6 +1403,30 @@ type JobsClientGetOptions struct {
 // JobsClientListOptions contains the optional parameters for the JobsClient.List method.
 type JobsClientListOptions struct {
 	// placeholder for future optional parameters
+}
+
+// KubernetesPVRestoreCriteria - Item Level kubernetes persistent volume target info for restore operation
+type KubernetesPVRestoreCriteria struct {
+	// REQUIRED; Type of the specific object - used for deserializing
+	ObjectType *string `json:"objectType,omitempty"`
+
+	// Selected persistent volume claim name
+	Name *string `json:"name,omitempty"`
+
+	// Selected storage class name for restore operation
+	StorageClassName *string `json:"storageClassName,omitempty"`
+}
+
+// KubernetesStorageClassRestoreCriteria - Item Level kubernetes storage class target info for restore operation
+type KubernetesStorageClassRestoreCriteria struct {
+	// REQUIRED; Type of the specific object - used for deserializing
+	ObjectType *string `json:"objectType,omitempty"`
+
+	// Provisioner of the storage class
+	Provisioner *string `json:"provisioner,omitempty"`
+
+	// Selected storage class name
+	SelectedStorageClassName *string `json:"selectedStorageClassName,omitempty"`
 }
 
 // MonitoringSettings - Monitoring Settings
@@ -1583,6 +1607,9 @@ type RecoveryPointsFilters struct {
 }
 
 type ResourceGuard struct {
+	// List of critical operations which are not protected by this resourceGuard
+	VaultCriticalOperationExclusionList []*string `json:"vaultCriticalOperationExclusionList,omitempty"`
+
 	// READ-ONLY; This flag indicates whether auto approval is allowed or not.
 	AllowAutoApprovals *bool `json:"allowAutoApprovals,omitempty" azure:"ro"`
 
@@ -1594,9 +1621,6 @@ type ResourceGuard struct {
 
 	// READ-ONLY; {readonly} List of operation details those are protected by the ResourceGuard resource
 	ResourceGuardOperations []*ResourceGuardOperation `json:"resourceGuardOperations,omitempty" azure:"ro"`
-
-	// READ-ONLY; List of critical operations which are not protected by this resourceGuard
-	VaultCriticalOperationExclusionList []*string `json:"vaultCriticalOperationExclusionList,omitempty" azure:"ro"`
 }
 
 // ResourceGuardOperation - This class contains all the details about a critical operation.